+++ src/framebuffer.h
@@ -1,24 +1,23 @@
 #pragma once
-#include <stddef.h>
 
 namespace texture {
 
 	void initFramebuffers();
 
 	void initGBuffer();
-	extern unsigned int gBuffer;
-	extern unsigned int gPosTexture;
-	extern unsigned int gNormalTexture;
-	extern unsigned int gAmbientTexture;
-	extern unsigned int gDiffuseTexture;
-	extern unsigned int gSpecularTexture;
-	extern unsigned int gDepthRenderbuffer;
+	extern size_t gBuffer;
+	extern size_t gPosTexture;
+	extern size_t gNormalTexture;
+	extern size_t gAmbientTexture;
+	extern size_t gDiffuseTexture;
+	extern size_t gSpecularTexture;
+	extern size_t gDepthRenderbuffer;
 
 	void initGUIFBO();
-	extern unsigned int guiFBO;
-	extern unsigned int quadIndexBuffer;
-	extern unsigned int quadIndexTexture;
-	extern unsigned int fontColorTexture;
-	extern unsigned int guiDepthRenderbuffer;
+	extern size_t guiFBO;
+	extern size_t quadIndexBuffer;
+	extern size_t quadIndexTexture;
+	extern size_t fontColorTexture;
+	extern size_t guiDepthRenderbuffer;
 }
